backend/config.py
from dotenv import load_dotenv
import os

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
SPOONACULAR_API_KEY = os.getenv("SPOONACULAR_API_KEY")
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///chat_history.db")
