requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
langchain==0.1.10
langgraph==0.0.20
langchain_openai==0.0.8
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
